requirements.txt
Flask==3.0.3
Requests==2.32.3
pyyaml==6.0.2